Ensure consistent HAPI-FHIR versions
While testing the Directory-Sync functionality, bugs started to occur
when the following call was made in DirectorySyncJob.createFhirApi():

client.registerInterceptor(new LoggingInterceptor(true));

Specifically, this was throwing an AbstractMethodError. It transpired
that this was due to inconsistent HAPI-FHIR versions, we were using a
mixture of 4.1.0 and 4.2.0.

This patch fixes that problem.