Fix Unexpected Behaviour of Count Obfuscation
First we now return zero when the actual value is zero in order to keep
unnecessary workloads away from the sites.

Second, we changed the epsilon from 0.12 to 0.28 in order to lower the
standard derivation from about 11 to about 5 in order to get back to
about the same variation as the old algorithm.

Third, we changed the sensitivity of the sample count to 10, because
most patients will have about 10 samples.